+++ packages/telemetry/browser-telemetry/src/collectors/rrweb/SessionReplay.ts
@@ -1,0 +1,39 @@
+/* eslint-disable max-classes-per-file */
+import * as rrweb from 'rrweb';
+
+import { BrowserTelemetry } from '../../api/BrowserTelemetry';
+import { Collector } from '../../api/Collector';
+import applyPatches from './patches';
+import SessionBuffer from './SessionBuffer';
+
+export default class SessionReplay implements Collector {
+  telemetry?: BrowserTelemetry;
+  buffer: SessionBuffer;
+
+  constructor(checkoutEveryNth: number, numBuffers: number) {
+    applyPatches();
+    this.buffer = new SessionBuffer(checkoutEveryNth, numBuffers);
+    const { buffer } = this;
+    rrweb.record({
+      checkoutEveryNth,
+      emit(event) {
+        buffer.push(event);
+      },
+    });
+  }
+
+  register(telemetry: BrowserTelemetry): void {
+    this.telemetry = telemetry;
+  }
+
+  unregister(): void {
+    this.telemetry = undefined;
+  }
+
+  capture(): void {
+    this.telemetry?.captureSession({
+      id: crypto.randomUUID(),
+      events: this.buffer.toArray(),
+    });
+  }
+}
