+++ packages/telemetry/browser-telemetry/src/index.ts
@@ -1,11 +1,10 @@
 import { BrowserTelemetry } from './api/BrowserTelemetry';
 import { Options } from './api/Options';
 import BrowserTelemetryImpl from './BrowserTelemetryImpl';
-import Session from './collectors/rrweb';
+import SessionReplay from './collectors/rrweb/SessionReplay';
 import parse from './options';
-import './patches';
 
-export { Session };
+export { SessionReplay };
 export { type Breadcrumb } from './api/Breadcrumb';
 export { type BrowserTelemetry } from './api/BrowserTelemetry';
 export { type Options } from './api/Options';
